Increases chances of TCGPlayer search result matching
Now it doesn't need to check for an English version of the card - if
there aren't multiple conditions, just accept the single result.

It also will compare against the collector number and set number for
cases of reprints or multiple copies of the same name in the same set,
e.g. Guildgates.