Switches how images from Scryfall are loaded
Hardcoding a URL stopped working due to a change in newer URLs. For
now I've changed it to store the URL for each card in the database.

Ideally it should be changed to dynamically grab the URL from
Scryfall, but there's a pretty huge impact on load times. Some form
of caching might be the way to go.