Deck management improvements
Importing can now be done. Currently working with expected format from
deckbox.org.

The Decks page will fetch a list of any decks in the database for the
current user, and display them in a sidenav list. I'm not 100% sold on
this design yet, it feels a little strange for the functionality. The
list will show a circular icon of a card in the deck, currently
defaulting to the first card in the deck.